# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "lafia nodejs server api",
    "version": "v1",
    "title": "LAFIA API"
  },
  "host": "127.0.0.1:7000",
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "scheme": "bearer",
      "in": "header",
      "description": "Authentication token"
    }
  },
  "paths": {
...
    "/user/modify/{uuid}": {
      "put": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Edit user by uuid",
        "description": "",
        "operationId": "updateuser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "user uuid need to be add in the url request",
            "required": true,
            "type": "string",
            "items": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "body",
            "description": "body for your query",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateusers"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "items": {
                "$ref": "#/definitions/getuser"
              }
            }
          },
          "401": {
            "description": "Invalid status value"
          }
        }
      }
    },
...
    "/company/{company_uuid}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get company by uuid",
        "description": "",
        "operationId": "getcompany",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "company_uuid",
            "in": "path",
            "description": "company uuid need to be add in the url request",
            "required": true,
            "type": "string",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "items": {
                "$ref": "#/definitions/getcompany"
              }
            }
          },
          "401": {
            "description": "Invalid status value"
          }
        }
      }
    },
...
    "/product/{product_uuid}": {
      "get": {
        "tags": [
          "Products"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get product by uuid",
        "description": "",
        "operationId": "getproduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "product_uuid",
            "in": "path",
            "description": "product uuid need to be add in the url request",
            "required": true,
            "type": "string",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "items": {
                "$ref": "#/definitions/getproduct"
              }
            }
          },
          "401": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/product/modify/{product_uuid}": {
      "put": {
        "tags": [
          "Products"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update product",
        "description": "",
        "operationId": "updateproduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "body for your query",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateproduct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "items": {
                "$ref": "#/definitions/getproduct"
              }
            }
          },
          "401": {
            "description": "Invalid status value"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "getuser": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "example": 201
        },
        "message": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "array",
          "example": {
            "_id": "t7457457t5*****",
            "firstName": "John",
            "lastName": "Doe",
            "address": "Bamako, Nord-ouest, rue 23",
            "email": "jer*****@gmail.com",
            "password": "$2b*******",
            "phone": "070*********",
            "role": "user",
            "referalCode": "***RT",
            "secretCode": null,
            "referrer": "445GH",
            "createdAt": "2022-11-29T14:32:04.681Z",
            "__v": 0
          }
        }
      }
    },
...
    "updateusers": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "address": {
          "type": "string",
          "example": "Bamako, rue 22"
        },
        "email": {
          "type": "string",
          "example": "email@example.com"
        },
        "phone": {
          "type": "string",
          "example": "+00 0000 000"
        },
        "referrer": {
          "type": "string",
          "example": "54FG"
        }
      }
    },
...
    "getcompany": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "example": 201
        },
        "message": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "example": {
            "_id": "t7457457t5*****",
            "name": "Lafia boutik",
            "description": "Description ...",
            "email":"jer*****@gmail.com",
            "phone":"000000000000",
            "location": "Bamako rue 22",
            "country": "Mali",
            "city": "Bamako",
            "departement": "ville basse",
            "categoryCompanyId": "t7457457t5*****",
            "status": "pending",
            "website":"www.me.com",
            "facebook":"www.me.com",
            "instagram":"www.me.com",
            "twitter":"www.me.com",
            "linkedin":"www.me.com",
            "createdAt": "2022-11-29T14:32:04.681Z",
            "updatedAt": "2022-11-29T14:32:04.681Z",
            "__v": 0
          }
        }
      }
    },
...
    "getproduct": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "example": 201
        },
        "message": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "example": {
            "_id": "63bd9a7a3dbbd70a6b131f55",
            "name": "T-shirt",
            "description": "Men clothes",
            "images": [
              "https://image-01-product-001.png"
            ],
            "price": 1000,
            "deliveryCost": 1000,
            "makingTime": "1h30",
            "categoryProductId": "63bc353ceb248d846835d669",
            "isAvailable" : true,
            "createdAt": "2023-01-10T17:03:54.224Z",
            "__v": 0
          }
        }
      }
    },
    "updateproduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Essentials size xl"
        },
        "description": {
          "type": "string",
          "example": "This is T-shirt for men"
        },
        "price": {
          "type": "number",
          "example": "1000"
        },
        "deliveryCost": {
          "type": "number",
          "example": "1000"
        },
        "makingTime": {
          "type": "string",
          "example": "1h30"
        },
        "categoryProductId": {
          "type": "string",
          "example": "12ehj990********"
        },
        "isAvalaible" : {
          "type":"boolean",
          "example": true
        },
        "userId": {
          "type": "string",
          "example": "12ehj990********"
        }
      }
    },
...
  },
...
}